Embed the Manual Client within the apworld
